+++ packages/core/android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -12,6 +12,7 @@ import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
 import android.net.Uri;
 import android.util.SparseIntArray;
+import androidx.annotation.VisibleForTesting;
 import androidx.core.app.FrameMetricsAggregator;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
@@ -19,6 +20,7 @@ import com.facebook.hermes.instrumentation.HermesSamplingProfiler;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.ReadableMapKeySetIterator;
 import com.facebook.react.bridge.ReadableType;
@@ -28,7 +30,6 @@ import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeArray;
 import com.facebook.react.bridge.WritableNativeMap;
 import com.facebook.react.common.JavascriptException;
-import com.facebook.react.modules.core.DeviceEventManagerModule;
 import io.sentry.Breadcrumb;
 import io.sentry.HubAdapter;
 import io.sentry.ILogger;
@@ -105,7 +106,7 @@ public class RNSentryModuleImpl {
   private FrameMetricsAggregator frameMetricsAggregator = null;
   private boolean androidXAvailable;
 
-  private static boolean hasFetchedAppStart;
+  @VisibleForTesting static long lastStartTimestampMs = -1;
 
   // 700ms to constitute frozen frames.
   private static final int FROZEN_FRAME_THRESHOLD = 700;
@@ -153,11 +154,7 @@ public class RNSentryModuleImpl {
   private @NotNull Runnable createEmitNewFrameEvent() {
     return () -> {
       final SentryDate endDate = dateProvider.now();
-      WritableMap event = Arguments.createMap();
-      event.putDouble("newFrameTimestampInSeconds", endDate.nanoTimestamp() / 1e9);
-      getReactApplicationContext()
-          .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
-          .emit("rn_sentry_new_frame", event);
+      RNSentryTimeToDisplay.putTimeToInitialDisplayForActiveSpan(endDate.nanoTimestamp() / 1e9);
     };
   }
 
@@ -432,31 +429,44 @@ public class RNSentryModuleImpl {
 
   public void fetchNativeAppStart(Promise promise) {
     fetchNativeAppStart(
-        promise,
-        InternalSentrySdk.getAppStartMeasurement(),
-        logger,
-        AppStartMetrics.getInstance().isAppLaunchedInForeground());
+        promise, AppStartMetrics.getInstance(), InternalSentrySdk.getAppStartMeasurement(), logger);
   }
 
   protected void fetchNativeAppStart(
       Promise promise,
-      final Map<String, Object> appStartMeasurement,
-      ILogger logger,
-      boolean isAppLaunchedInForeground) {
-    if (!isAppLaunchedInForeground) {
+      final AppStartMetrics metrics,
+      final Map<String, Object> metricsDataBag,
+      ILogger logger) {
+    if (!metrics.isAppLaunchedInForeground()) {
       logger.log(SentryLevel.WARNING, "Invalid app start data: app not launched in foreground.");
       promise.resolve(null);
       return;
     }
 
     WritableMap mutableMeasurement =
-        (WritableMap) RNSentryMapConverter.convertToWritable(appStartMeasurement);
-    mutableMeasurement.putBoolean("has_fetched", hasFetchedAppStart);
+        (WritableMap) RNSentryMapConverter.convertToWritable(metricsDataBag);
 
-    // This is always set to true, as we would only allow an app start fetch to only
-    // happen once in the case of a JS bundle reload, we do not want it to be
-    // instrumented again.
-    hasFetchedAppStart = true;
+    long currentStartTimestampMs = metrics.getAppStartTimeSpan().getStartTimestampMs();
+    boolean hasFetched =
+        lastStartTimestampMs > 0 && lastStartTimestampMs == currentStartTimestampMs;
+    mutableMeasurement.putBoolean("has_fetched", hasFetched);
+
+    if (lastStartTimestampMs < 0) {
+      logger.log(SentryLevel.DEBUG, "App Start data reported to the RN layer for the first time.");
+    } else if (hasFetched) {
+      logger.log(SentryLevel.DEBUG, "App Start data already fetched from native before.");
+    } else {
+      logger.log(SentryLevel.DEBUG, "App Start data updated, reporting to the RN layer again.");
+    }
+
+    // When activity is destroyed but the application process is kept alive
+    // the next activity creation is considered warm start.
+    // The app start metrics will be updated by the the Android SDK.
+    // To let the RN JS layer know these are new start data we compare the start timestamps.
+    lastStartTimestampMs = currentStartTimestampMs;
+
+    // Clears start metrics, making them ready for recording warm app start
+    metrics.onAppStartSpansSent();
 
     promise.resolve(mutableMeasurement);
   }
@@ -701,6 +711,19 @@ public class RNSentryModuleImpl {
         scope -> {
           scope.clearBreadcrumbs();
         });
+  }
+
+  public void popTimeToDisplayFor(String screenId, Promise promise) {
+    if (screenId != null) {
+      promise.resolve(RNSentryTimeToDisplay.popTimeToDisplayFor(screenId));
+    } else {
+      promise.resolve(null);
+    }
+  }
+
+  public boolean setActiveSpanId(@Nullable String spanId) {
+    RNSentryTimeToDisplay.setActiveSpanId(spanId);
+    return true; // The return ensure RN executes the code synchronously
   }
 
   public void setExtra(String key, String extra) {
@@ -1003,6 +1026,15 @@ public class RNSentryModuleImpl {
       logger.log(SentryLevel.ERROR, msg);
       promise.reject(new Exception(msg));
     }
+  }
+
+  public void encodeToBase64(ReadableArray array, Promise promise) {
+    byte[] bytes = new byte[array.size()];
+    for (int i = 0; i < array.size(); i++) {
+      bytes[i] = (byte) array.getInt(i);
+    }
+    String base64String = android.util.Base64.encodeToString(bytes, android.util.Base64.DEFAULT);
+    promise.resolve(base64String);
   }
 
   public void crashedLastRun(Promise promise) {