+++ packages/core/android/src/oldarch/java/io/sentry/react/RNSentryModule.java
@@ -157,6 +157,11 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
     this.impl.getDataFromUri(uri, promise);
   }
 
+  @ReactMethod
+  public void encodeToBase64(ReadableArray array, Promise promise) {
+    this.impl.encodeToBase64(array, promise);
+  }
+
   @ReactMethod(isBlockingSynchronousMethod = true)
   public WritableMap fetchNativeStackFramesBy(ReadableArray instructionsAddr) {
     // Not used on Android
@@ -181,5 +186,10 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
   @ReactMethod()
   public void getNewScreenTimeToDisplay(Promise promise) {
     this.impl.getNewScreenTimeToDisplay(promise);
+  }
+
+  @ReactMethod
+  public boolean setActiveSpanId(String spanId) {
+    return this.impl.setActiveSpanId(spanId);
   }
 }