+++ packages/core/src/js/feedback/defaults.ts
@@ -1,4 +1,4 @@
-import type { FeedbackWidgetProps } from './FeedbackWidget.types';
+import type { FeedbackButtonProps, FeedbackWidgetProps, ScreenshotButtonProps } from './FeedbackWidget.types';
 import { feedbackAlertDialog } from './utils';
 
 const FORM_TITLE = 'Report a Bug';
@@ -11,11 +11,14 @@ const MESSAGE_LABEL = 'Description';
 const IS_REQUIRED_LABEL = '(required)';
 const SUBMIT_BUTTON_LABEL = 'Send Bug Report';
 const CANCEL_BUTTON_LABEL = 'Cancel';
+const TRIGGER_LABEL = 'Report a Bug';
+const TRIGGER_SCREENSHOT_LABEL = 'Take Screenshot';
 const ERROR_TITLE = 'Error';
 const FORM_ERROR = 'Please fill out all required fields.';
 const EMAIL_ERROR = 'Please enter a valid email address.';
 const SUCCESS_MESSAGE_TEXT = 'Thank you for your report!';
 const ADD_SCREENSHOT_LABEL = 'Add a screenshot';
+const CAPTURE_SCREENSHOT_LABEL = 'Take a screenshot';
 const REMOVE_SCREENSHOT_LABEL = 'Remove screenshot';
 const GENERIC_ERROR_TEXT = 'Unable to send feedback due to an unexpected error.';
 
@@ -60,6 +63,7 @@ export const defaultConfiguration: Partial<FeedbackWidgetProps> = {
   showEmail: true,
   showName: true,
   enableScreenshot: false,
+  enableTakeScreenshot: false,
 
   // FeedbackTextConfiguration
   cancelButtonLabel: CANCEL_BUTTON_LABEL,
@@ -78,5 +82,16 @@ export const defaultConfiguration: Partial<FeedbackWidgetProps> = {
   successMessageText: SUCCESS_MESSAGE_TEXT,
   addScreenshotButtonLabel: ADD_SCREENSHOT_LABEL,
   removeScreenshotButtonLabel: REMOVE_SCREENSHOT_LABEL,
+  captureScreenshotButtonLabel: CAPTURE_SCREENSHOT_LABEL,
   genericError: GENERIC_ERROR_TEXT,
+};
+
+export const defaultButtonConfiguration: Partial<FeedbackButtonProps> = {
+  triggerLabel: TRIGGER_LABEL,
+  triggerAriaLabel: '',
+};
+
+export const defaultScreenshotButtonConfiguration: Partial<ScreenshotButtonProps> = {
+  triggerLabel: TRIGGER_SCREENSHOT_LABEL,
+  triggerAriaLabel: '',
 };