+++ packages/core/src/js/index.ts
@@ -79,13 +79,16 @@ export {
   startIdleNavigationSpan,
   startIdleSpan,
   getDefaultIdleNavigationSpanOptions,
+  createTimeToFullDisplay,
+  createTimeToInitialDisplay,
 } from './tracing';
 
 export type { TimeToDisplayProps } from './tracing';
 
 export { Mask, Unmask } from './replay/CustomMask';
 
+export { FeedbackButton } from './feedback/FeedbackButton';
 export { FeedbackWidget } from './feedback/FeedbackWidget';
-export { showFeedbackWidget } from './feedback/FeedbackWidgetManager';
+export { showFeedbackWidget, showFeedbackButton, hideFeedbackButton } from './feedback/FeedbackWidgetManager';
 
 export { getDataFromUri } from './wrapper';