+++ packages/core/src/js/wrapper.ts
@@ -122,6 +122,11 @@ interface SentryNativeWrapper {
   getNewScreenTimeToDisplay(): Promise<number | null | undefined>;
 
   getDataFromUri(uri: string): Promise<Uint8Array | null>;
+  popTimeToDisplayFor(key: string): Promise<number | undefined | null>;
+
+  setActiveSpanId(spanId: string): void;
+
+  encodeToBase64(data: Uint8Array): Promise<string | null>;
 }
 
 const EOL = utf8ToBytes('\n');
@@ -714,6 +719,47 @@ export const NATIVE: SentryNativeWrapper = {
     } catch (error) {
       logger.error('Error:', error);
       return null;
+    }
+  },
+
+  popTimeToDisplayFor(key: string): Promise<number | undefined | null> {
+    if (!this.enableNative || !this._isModuleLoaded(RNSentry)) {
+      return Promise.resolve(null);
+    }
+
+    try {
+      return RNSentry.popTimeToDisplayFor(key);
+    } catch (error) {
+      logger.error('Error:', error);
+      return null;
+    }
+  },
+
+  setActiveSpanId(spanId): void {
+    if (!this.enableNative || !this._isModuleLoaded(RNSentry)) {
+      return undefined;
+    }
+
+    try {
+      RNSentry.setActiveSpanId(spanId);
+    } catch (error) {
+      logger.error('Error:', error);
+      return undefined;
+    }
+  },
+
+  async encodeToBase64(data: Uint8Array): Promise<string | null> {
+    if (!this.enableNative || !this._isModuleLoaded(RNSentry)) {
+      return Promise.resolve(null);
+    }
+
+    try {
+      const byteArray = Array.from(data);
+      const base64 = await RNSentry.encodeToBase64(byteArray);
+      return base64 || null;
+    } catch (error) {
+      logger.error('Error:', error);
+      return Promise.resolve(null);
     }
   },