+++ samples/react-native/e2e/captureErrorsScreenTransaction.test.ts
@@ -7,29 +7,22 @@ import {
 
 import { getItemOfTypeFrom } from './utils/event';
 import { maestro } from './utils/maestro';
-import { isAndroid, isIOS } from './utils/environment';
 
-describe('Capture transaction', () => {
+describe('Capture Errors Screen Transaction', () => {
   let sentryServer = createSentryServer();
   sentryServer.start();
 
   const getErrorsEnvelope = () =>
     sentryServer.getEnvelope(containingTransactionWithName('Errors'));
 
-  const getTrackerEnvelope = () =>
-    sentryServer.getEnvelope(containingTransactionWithName('Tracker'));
-
   beforeAll(async () => {
-    const waitForTrackerTx = sentryServer.waitForEnvelope(
-      containingTransactionWithName('Tracker'), // The last created and sent transaction
-    );
     const waitForErrorsTx = sentryServer.waitForEnvelope(
       containingTransactionWithName('Errors'), // The last created and sent transaction
     );
 
-    await maestro('captureTransaction.test.yml');
+    await maestro('captureErrorsScreenTransaction.test.yml');
 
-    await Promise.all([waitForTrackerTx, waitForErrorsTx]);
+    await waitForErrorsTx;
   });
 
   afterAll(async () => {
@@ -79,38 +72,20 @@ describe('Capture transaction', () => {
       'transaction',
     );
 
-    if (isIOS()) {
-      expect(item?.[1]).toEqual(
-        expect.objectContaining({
-          measurements: expect.objectContaining({
-            time_to_initial_display: {
-              unit: 'millisecond',
-              value: expect.any(Number),
-            },
-            app_start_cold: {
-              unit: 'millisecond',
-              value: expect.any(Number),
-            },
-          }),
+    expect(item?.[1]).toEqual(
+      expect.objectContaining({
+        measurements: expect.objectContaining({
+          time_to_initial_display: {
+            unit: 'millisecond',
+            value: expect.any(Number),
+          },
+          app_start_cold: {
+            unit: 'millisecond',
+            value: expect.any(Number),
+          },
         }),
-      );
-    } else if (isAndroid()) {
-      // TMP: Until the cold app start is fixed on Android
-      expect(item?.[1]).toEqual(
-        expect.objectContaining({
-          measurements: expect.objectContaining({
-            time_to_initial_display: {
-              unit: 'millisecond',
-              value: expect.any(Number),
-            },
-            app_start_warm: {
-              unit: 'millisecond',
-              value: expect.any(Number),
-            },
-          }),
-        }),
-      );
-    }
+      }),
+    );
   });
 
   it('contains time to initial display measurements', async () => {
@@ -153,56 +128,6 @@ describe('Capture transaction', () => {
             value: expect.any(Number),
           },
         }),
-      }),
-    );
-  });
-
-  it('contains time to display measurements', async () => {
-    const item = getItemOfTypeFrom<EventItem>(
-      getTrackerEnvelope(),
-      'transaction',
-    );
-
-    expect(item?.[1]).toEqual(
-      expect.objectContaining({
-        measurements: expect.objectContaining({
-          time_to_initial_display: {
-            unit: 'millisecond',
-            value: expect.any(Number),
-          },
-          time_to_full_display: {
-            unit: 'millisecond',
-            value: expect.any(Number),
-          },
-        }),
-      }),
-    );
-  });
-
-  it('contains at least one xhr breadcrumb of request to the tracker endpoint', async () => {
-    const item = getItemOfTypeFrom<EventItem>(
-      getTrackerEnvelope(),
-      'transaction',
-    );
-
-    expect(item?.[1]).toEqual(
-      expect.objectContaining({
-        breadcrumbs: expect.arrayContaining([
-          expect.objectContaining({
-            category: 'xhr',
-            data: {
-              end_timestamp: expect.any(Number),
-              method: 'GET',
-              response_body_size: expect.any(Number),
-              start_timestamp: expect.any(Number),
-              status_code: expect.any(Number),
-              url: expect.stringContaining('api.covid19api.com/summary'),
-            },
-            level: 'info',
-            timestamp: expect.any(Number),
-            type: 'http',
-          }),
-        ]),
       }),
     );
   });